requirements.txt
pika
flask
mysql-connector-python